requirements.txt
aiogram==3.13.1
YooMoney==0.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
